backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cookbook"
version = "1.0.0"
description = "Modern recipe management system with FastAPI backend and React frontend"
authors = [{name = "Anton Lu", email = "hello@antonlu.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi~=0.104.1",
    "uvicorn[standard]~=0.24.0",
    "sqlalchemy~=2.0.23",
    "alembic~=1.13.0",
    "asyncpg~=0.29.0",
    "redis~=5.0.1",
    "celery~=5.3.4",
    "pillow~=10.1.0",
    "python-multipart~=0.0.6",
    "passlib[bcrypt]~=1.7.4",
    "python-jose[cryptography]~=3.3.0",
    "pydantic~=2.5.0",
    "pydantic-settings~=2.1.0",
    "httpx~=0.25.2",
    "pyyaml~=6.0.1",
    "markdown~=3.5.1",
    "pymdown-extensions~=10.4",
    "python-frontmatter~=1.1.0",
    "fastapi-users[sqlalchemy]~=12.1.2",
    "pyjwt~=2.8.0",
]

[project.optional-dependencies]
test = [
    "pytest~=7.4.3",
    "pytest-asyncio~=0.21.1",
    "pytest-cov~=4.1.0",
    "pytest-mock~=3.12.0",
    "pytest-benchmark~=4.0.0",
    "pytest-xdist~=3.5.0",
    "fakeredis~=2.20.1",
    "factory-boy~=3.3.0",
    "freezegun~=1.2.2",
    "responses~=0.24.1",
    "httpx~=0.25.2",
    "aiofiles~=23.2.1",
    "aiosqlite~=0.19.0",
    "sqlalchemy-utils~=0.41.1",
    "pytest-clarity~=1.0.1",
]
security = [
    "bandit~=1.7.5",
    "safety~=2.3.5",
]
dev = [
    "black~=23.11.0",
    "ruff~=0.1.6",
    "mypy~=1.7.1",
    "pre-commit~=3.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=html",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "security: Security tests",
    "performance: Performance tests",
    "slow: Slow running tests",
    "auth: Authentication related tests",
    "upload: File upload related tests",
    "database: Database related tests",
    "recipe: Recipe management tests",
    "markdown: Markdown processing tests",
]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]

exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  ".venv",
]

line-length = 88
target-version = "py311"

preview = true
unsafe-fixes = true

[tool.ruff.lint]
# All rules can be found here: https://docs.astral.sh/ruff/rules/
select = [
  "F",   # pyflakes
  "E",   # error  (pycodestyle)
  "W",   # warning (pycodestyle)
  "I",   # isort
  "N",   # naming  (pep8)
  "UP",  # pyupgrade
  "PLC", # convention  (pylint)
  "PLE", # error  (pylint)
  "PLW", # warning  (pylint)
  # "D",  # docstring (pydocstyle),  # not enabled by default
  "PD",   # pandas
  "TRY",  # tryceratops
  "NPY",  # numpy
  "PERF", # perflint
  "RUF",  # ruff
  # v flake8 v
  "B",
  "C4",
  "FBT",
  "A",
  "EM",
  "ISC",
  "FA",
  "G",
  "PIE",
  "PYI",
  "PT",
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SIM",
]
fixable = ["ALL"]
ignore = [
   "E501",
   "W505",
   "ISC001",
   "PD901",
   "PLW2901",
   "N812",
   "N806",
   "G004",
   "RUF002",
   "FBT001",   # Boolean-typed positional argument
   "FBT002",   # Boolean default positional argument
   "FBT003",   # Boolean positional value in function call
   "NPY002",   # Replace legacy numpy random calls
   "E721",     # Type comparison with ==
   "RUF043",   # Raw string for regex patterns
   "B008",     # Do not perform function calls in argument defaults (FastAPI Depends pattern)
   "TRY300",   # Consider moving statement to else block (stylistic)
   "TRY401",   # Redundant exception object in logging.exception
   "RUF029",   # Function is declared async but doesn't await (acceptable for FastAPI)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "SLF001",    # Private member accessed (common in tests)
  "PLC2701",   # Private name import (common in tests)
  "PLC0415",   # Import should be at top-level (common in test functions)
  "PT011",     # pytest.raises() too broad (acceptable in tests)
  "PT017",     # assertion on exception in except block (acceptable in tests)
  "PT012",     # pytest.raises() multi-statement (acceptable in tests)
  "SIM115",    # Use context manager for opening files (acceptable in tests)
  "TRY300",    # Move statement to else block (acceptable in tests)
  "PLC1901",   # Empty string comparison (acceptable in tests)
]
"docs/**" = [
  "SLF001",    # Private member accessed (needed for docs)
]
"app/core/image_processor.py" = [
  "SLF001",    # Private member accessed (_getexif is the correct PIL API)
]
"test_simple.py" = [
  "PLC0415",   # Import at function level (demo script)
]

[tool.ruff.format]
preview = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_errors = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.crud.*"
warn_return_any = false

[tool.bandit]
exclude_dirs = ["tests", "migrations", "test_simple.py"]
skips = ["B101", "B601", "B105"]  # B105: hardcoded password string (common in tests)